change samesite setting on cookie created on service
when cookie is secure: false and none it will be discarded by browser anyway, so if its not secure set it to lax. made to allow to test on dev when all is http so it should be secure:false and on same site. and when on prod sites are probably different, but then it will use https